Use an `APIRouter` to separate text and voice endpoints.
